Isolate remembered home link component
